TEMPORARY / LPS-73731 Set servlet context in each dynamic include.
Once we have a reference set by a public/protected method in a superclass, it
is not possible to override the setter in subclasses, because that would result
in multiple refs with same name set. We need to override this set when creating
dynamic includes that use other servlet contexts (otherwise we would not be
able to include JSPs from other modules) so we have no option other than
imlementing setServletContext() in every class at the leaves of the inheritance
tree.